+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added support for setting a custom transport adapter by implementing the method `std::shared_ptr<HttpTransport> ::AzureSdkGetCustomHttpTransport()`.
 - Added default constructor to `DateTime`.
 - Added Base64 encoding and decoding utility APIs to the `Azure::Core` namespace available from `azure/core/base64.hpp`.
+- Added `Http::Response<void>` template specialization.
 
 ### Breaking Changes