+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.4 (2021-1-11)
+## 1.0.0-beta.4 (2021-1-12)
 
 ### New Features